requirements.txt
PyGithub==1.46
requests==2.23.0